requirements.txt
opencv-python>=4.8.0
numpy>=1.24.0
Pillow>=9.0.0
python-dotenv>=1.0.0
openai>=1.0.0
azure-core>=1.26.0
azure-ai-formrecognizer>=3.2.0
rich>=13.7.0
matplotlib>=3.7.0
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart>=0.0.5
pydantic>=1.8.2
